refactor: Constant, ErrorMessage 패키지 이동
